requirements.txt
fastapi==0.108.0
firebase-admin==6.1.0
gnupg==2.3.1
slowapi==0.1.8
h11==0.14.0
httptools==0.5.0
pydantic==1.10.2
PyJWT==2.6.0
pyodbc==4.0.34
python-dotenv==0.21.0
python-multipart==0.0.6
requests==2.31.0
rsa==4.9
starlette==0.30.0
typing_extensions==4.9.0
urllib3==1.26.18
utils==1.0.1
uvicorn==0.19.0
watchfiles==0.18.1
